refactor(ui): split LogDetailModal into header/body/raw sub-components
